requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
fastjsonschema==2.20.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
ipython==8.12.3
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyterlab_pygments==0.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
narwhals==1.6.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.37.1
pandas==2.2.2
pandocfilters==1.5.1
pillow==10.4.0
pinecone==5.1.0
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
pipreqs==0.5.0
protobuf==5.28.0
psycopg2-binary==2.9.10
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
setuptools==71.0.1
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
streamlit==1.38.0
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.19.1
toml==0.10.2
torch==2.3.1
tqdm==4.66.4
transformers==4.42.4
tzdata==2024.1
urllib3==2.2.2
uuid==1.30
webencodings==0.5.1
wheel==0.43.0
wrapt==1.16.0
yarg==0.1.9
yarl==1.9.4
